+++ R/BASiCS-estimate.R
@@ -34,7 +34,7 @@
 #' @examples
 #' \dontrun{
 #' # Load example data
-#' library(scater)
+#' library(scuttle)
 #' set.seed(1)
 #' sce <- mockSCE()
 #'