+++ R/kersplat-estimate.R
@@ -19,7 +19,7 @@
 #'
 #' if (requireNamespace("igraph", quietly = TRUE)) {
 #'     # Load example data
-#'     library(scater)
+#'     library(scuttle)
 #'     set.seed(1)
 #'     sce <- mockSCE()
 #'