+++ R/lun-simulate.R
@@ -50,7 +50,7 @@ lunSimulate <- function(params = newLunParams(), sparsify = TRUE,
 
     # Set random seed
     seed <- getParam(params, "seed")
-    set.seed(seed)
+    withr::with_seed(seed, {
 
     # Get the parameters we are going to use
     nGenes <- getParam(params, "nGenes")
@@ -144,8 +144,8 @@ lunSimulate <- function(params = newLunParams(), sparsify = TRUE,
         assays(sim) <- sparsifyMatrices(assays(sim), auto = TRUE,
                                         verbose = verbose)
     }
+    })
 
     if (verbose) {message("Done!")}
-
     return(sim)
 }