+++ R/lun2-estimate.R
@@ -21,7 +21,7 @@
 #' @examples
 #' \dontrun{
 #' # Load example data
-#' library(scater)
+#' library(scuttle)
 #' set.seed(1)
 #' sce <- mockSCE()
 #'