+++ R/mfa-simulate.R
@@ -41,7 +41,7 @@ mfaSimulate <- function(params = newMFAParams(), sparsify = TRUE,
 
     # Set random seed
     seed <- getParam(params, "seed")
-    set.seed(seed)
+    withr::with_seed(seed, {
 
     # Get the parameters we are going to use
     nCells <- getParam(params, "nCells")
@@ -96,6 +96,7 @@ mfaSimulate <- function(params = newMFAParams(), sparsify = TRUE,
         assays(sim) <- sparsifyMatrices(assays(sim), auto = TRUE,
                                         verbose = verbose)
     }
+    })
 
     return(sim)
 }