+++ R/mock-data.R
@@ -4,7 +4,7 @@
 #'
 #' @param n.genes Number of genes in mock gff file
 #' @param chromosome Chromosome name
-#' @param chr.length Length of mock chromosome 
+#' @param chr.length Length of mock chromosome
 #' @param seed Optional: seed for random seed
 #'
 #' @return data.frame containing mock gff data.
@@ -14,14 +14,28 @@
 #'
 #' @export
 mockGFF <- function(n.genes = 50, chromosome = 1, chr.length = 2e6, seed = NULL){
-    if(!is.null(seed)){set.seed(seed)}
-    
-    mock.gff <- data.frame(list(V1 = chromosome,
-                                V2 = "source",
-                                V3 = "gene",
-                                V4 = sort(sample(1e4:chr.length, n.genes))))
-    mock.gff$V5 <- mock.gff$V4 + floor(rnorm(n.genes, 1500, 1000))
-    mock.gff[, c("V6", "V7", "V8", "V9")] <- "."
+
+    if (is.null(seed)) {
+        V4 = sort(sample(1e4:chr.length, n.genes))
+        V5 <- V4 + floor(rnorm(n.genes, 1500, 1000))
+    } else {
+        withr::with_seed(seed, {
+            V4 = sort(sample(1e4:chr.length, n.genes))
+            V5 <- V4 + floor(rnorm(n.genes, 1500, 1000))
+        })
+    }
+
+    mock.gff <- data.frame(
+        V1 = chromosome,
+        V2 = "source",
+        V3 = "gene",
+        V4 = V4,
+        V5 = V5,
+        V6 = ".",
+        V7 = ".",
+        V8 = ".",
+        V9 = "."
+    )
 
     return(mock.gff)
 }
@@ -35,7 +49,7 @@ mockGFF <- function(n.genes = 50, chromosome = 1, chr.length = 2e6, seed = NULL)
 #' @param n.snps Number of SNPs in mock vcf file.
 #' @param n.samples Number of samples in mock bulk data.
 #' @param chromosome Chromosome name
-#' @param chr.length Length of mock chromosome 
+#' @param chr.length Length of mock chromosome
 #' @param seed Optional: seed for random seed
 #'
 #' @return data.frame containing mock vcf data.
@@ -45,10 +59,9 @@ mockGFF <- function(n.genes = 50, chromosome = 1, chr.length = 2e6, seed = NULL)
 #'
 #' @export
 #' @importFrom stats setNames
-mockVCF <- function(n.snps = 200, n.samples = 5, 
+mockVCF <- function(n.snps = 200, n.samples = 5,
                     chromosome = 1,  chr.length = 2e6, seed = NULL){
 
-    if(!is.null(seed)){set.seed(seed)}
     checkDependencies(deps = "VariantAnnotation")
 
     sample_names <- paste0("sample_", formatC(seq_len(n.samples),
@@ -58,12 +71,23 @@ mockVCF <- function(n.snps = 200, n.samples = 5,
     snp_names <- paste0("snp_", formatC(seq_len(n.snps),
                                         width = nchar(n.snps),
                                         format = "d", flag = "0"))
+
+    if (is.null(seed)) {
+        ranges <- IRanges::IRanges(sample(seq_len(chr.length), n.snps,
+                                          replace = FALSE),
+                                   names = snp_names)
+    } else {
+        withr::with_seed(seed, {
+            ranges <- IRanges::IRanges(sample(seq_len(chr.length), n.snps,
+                                              replace = FALSE),
+                                       names = snp_names)
+        })
+    }
+
     # rowRanges
     vcf.rowRanges <- GenomicRanges::GRanges(
         seqnames = S4Vectors::Rle(rep(chromosome, n.snps)),
-        ranges = IRanges::IRanges(sample(seq_len(chr.length), n.snps, 
-                                         replace = FALSE),
-                                  names = snp_names),
+        ranges = ranges,
         strand = S4Vectors::Rle(BiocGenerics::strand(rep("*", n.snps))),
         paramRangeID = S4Vectors::Rle(rep(NA, n.snps))
     )
@@ -72,8 +96,17 @@ mockVCF <- function(n.snps = 200, n.samples = 5,
     genotypes <- c(rep("0|0", 100), rep("1|0", 10), rep("1|1", 10))
     geno <- setNames(data.frame(matrix(ncol = n.samples, nrow = n.snps)),
                      sample_names)
-    geno[,  sample_names] <- sample(genotypes, n.samples * n.snps,
-                                    replace = TRUE)
+
+    if (is.null(seed)) {
+        geno[,  sample_names] <- sample(genotypes, n.samples * n.snps,
+                                        replace = TRUE)
+    } else {
+        withr::with_seed(seed, {
+            geno[,  sample_names] <- sample(genotypes, n.samples * n.snps,
+                                            replace = TRUE)
+        })
+    }
+
     row.names(geno) <- snp_names
     geno <- as.matrix(geno)
 
@@ -119,8 +152,7 @@ mockVCF <- function(n.snps = 200, n.samples = 5,
 #'
 #' @export
 mockBulkMatrix <- function(n.genes = 100, n.samples = 50, seed = NULL){
-    
-    if(!is.null(seed)){set.seed(seed)}
+
     tmp.params <- newSplatPopParams()
     mean.shape <- getParam(tmp.params, "pop.mean.shape")
     mean.rate <- getParam(tmp.params, "pop.mean.rate")
@@ -128,14 +160,31 @@ mockBulkMatrix <- function(n.genes = 100, n.samples = 50, seed = NULL){
     cv.shape <- cv.df[5, "shape"]
     cv.rate <- cv.df[5, "rate"]
 
-    key <- data.frame(list(id = seq_len(n.genes),
-                           mean = rgamma(n.genes, mean.shape, mean.rate),
-                           cv = rgamma(n.genes, cv.shape, cv.rate)))
+    if (is.null(seed)) {
+        key <- data.frame(
+            id = seq_len(n.genes),
+            mean = rgamma(n.genes, mean.shape, mean.rate),
+            cv = rgamma(n.genes, cv.shape, cv.rate)
+        )
 
-    mock.means <- lapply(key$id, function(g) rnorm(n.samples,
-                                              mean = key[key$id == g,]$mean,
-                                              sd = key[key$id == g,]$mean *
-                                                  key[key$id == g,]$cv))
+        mock.means <- lapply(key$id, function(g) {
+            rnorm(n.samples, mean = key[key$id == g,]$mean,
+                  sd = key[key$id == g,]$mean * key[key$id == g,]$cv)
+        })
+    } else {
+        withr::with_seed(seed, {
+            key <- data.frame(
+                id = seq_len(n.genes),
+                mean = rgamma(n.genes, mean.shape, mean.rate),
+                cv = rgamma(n.genes, cv.shape, cv.rate)
+            )
+
+            mock.means <- lapply(key$id, function(g) {
+                rnorm(n.samples, mean = key[key$id == g,]$mean,
+                      sd = key[key$id == g,]$mean * key[key$id == g,]$cv)
+            })
+        })
+    }
 
     mock.means <- do.call(rbind, mock.means)
     mock.means[mock.means < 0] <- 0
@@ -158,67 +207,88 @@ mockBulkMatrix <- function(n.genes = 100, n.samples = 50, seed = NULL){
 #'
 #' @export
 mockBulkeQTL <- function(n.genes = 500, seed = NULL){
-    
-    if(!is.null(seed)){set.seed(seed)}
+
     tmp.params <- newSplatPopParams()
     eqtl.shape <- getParam(tmp.params, "eqtl.ES.shape")
     eqtl.rate <- getParam(tmp.params, "eqtl.ES.rate")
 
-    mock.eq <- data.frame(list(gene_id = seq_len(n.genes),
-                               pval_nominal = 0.01,
-                               slope = rgamma(n.genes, eqtl.shape, eqtl.rate)))
+    if (is.null(seed)) {
+        mock.eq <- data.frame(
+            gene_id = seq_len(n.genes),
+            pval_nominal = 0.01,
+            slope = rgamma(n.genes, eqtl.shape, eqtl.rate)
+        )
+    } else {
+        withr::with_seed(seed, {
+            mock.eq <- data.frame(
+                gene_id = seq_len(n.genes),
+                pval_nominal = 0.01,
+                slope = rgamma(n.genes, eqtl.shape, eqtl.rate)
+            )
+        })
+    }
 
     return(mock.eq)
 }
 
 
 #' Generate set of "empirical" mock data
-#' 
+#'
 #' Quick function to generate matching mock VCF, bulk expression, and eQTL data,
 #' useful for running splatPopEmpiricalMeans
-#' 
+#'
 #' @param n.genes Number of genes in mock eQTL data.
 #' @param n.snps Number of SNPs in mock vcf file.
 #' @param n.samples Number of samples in mock bulk data.
 #' @param chromosome Chromosome name
-#' @param chr.length Length of mock chromosome 
+#' @param chr.length Length of mock chromosome
 #' @param seed Optional: seed for random seed
-#' 
+#'
 #' @return list(gff=mockGFF, vcf=mockVCF, means=mockMEANS, eqtl=mockEQTL)
 #'
 #' @examples
 #' empirical <- mockEmpiricalSet()
 #'
 #' @export
-#' 
-mockEmpiricalSet <- function(n.genes = 20, n.snps = 1000, n.samples = 10, 
+#'
+mockEmpiricalSet <- function(n.genes = 20, n.snps = 1000, n.samples = 10,
                     chromosome = 1,  chr.length = 2e6, seed = NULL){
-    
-    mockGFF <- mockGFF(n.genes = n.genes, chromosome = chromosome, 
-                   chr.length = chr.length, seed=seed)
-    mockVCF <- mockVCF(n.snps = n.snps, n.samples = n.samples, 
-                   chromosome = chromosome,  chr.length = chr.length, seed=seed)
-    mockMEANS <- mockBulkMatrix(n.genes = n.genes, n.samples = n.samples, seed=seed)
-    mockEQTL <- mockBulkeQTL(n.genes = n.genes, seed=seed)
-    
-    mockEQTL$geneID <- paste0("gene_", formatC(seq_len(n.genes), 
+
+    mockGFF <- mockGFF(n.genes = n.genes, chromosome = chromosome,
+                   chr.length = chr.length, seed = seed)
+    mockVCF <- mockVCF(n.snps = n.snps, n.samples = n.samples,
+                       chromosome = chromosome,  chr.length = chr.length,
+                       seed = seed)
+    mockMEANS <- mockBulkMatrix(n.genes = n.genes, n.samples = n.samples,
+                                seed = seed)
+    mockEQTL <- mockBulkeQTL(n.genes = n.genes, seed = seed)
+
+    mockEQTL$geneID <- paste0("gene_", formatC(seq_len(n.genes),
                                                width = nchar(n.genes),
                                                format = "d", flag = "0"))
-    
+
     row.names(mockMEANS) <- mockEQTL$geneID
     colnames(mockMEANS) <- colnames(mockVCF)
-    
+
     vcfDF <- data.frame(SummarizedExperiment::rowRanges(mockVCF))
     row.names(vcfDF) <- rownames(mockVCF)
     vcfDF$MAF <- VariantAnnotation::snpSummary(mockVCF)$a1Freq
-    eSNPs <- sample(rownames(mockVCF), n.genes)
+
+    if (is.null(seed)) {
+        eSNPs <- sample(rownames(mockVCF), n.genes)
+    } else {
+        withr::with_seed(seed, {
+            eSNPs <- sample(rownames(mockVCF), n.genes)
+        })
+    }
+
     vcfDF <- vcfDF[eSNPs, ]
-    
+
     mockEQTL$snpID <- eSNPs
     mockEQTL$snpLOC <- vcfDF$start
     mockEQTL$snpCHR <- vcfDF$seqnames
     mockEQTL$snpMAF <- vcfDF$MAF
-    
+
     mockEQTL[mockEQTL$snpMAF < 0.05, c("snpID", "snpLOC", "snpCHR")] <- NA
 
     return(list(gff=mockGFF, vcf=mockVCF, means=mockMEANS, eqtl=mockEQTL))