+++ R/zinb-estimate.R
@@ -29,7 +29,7 @@
 #'
 #' @examples
 #' if (requireNamespace("zinbwave", quietly = TRUE)) {
-#'     library(scater)
+#'     library(scuttle)
 #'     set.seed(1)
 #'     sce <- mockSCE(ncells = 20, ngenes = 100)
 #'