+++ tests/testthat/test-lunEstimate.R
@@ -1,6 +1,6 @@
 context("lunEstimate")
 
-library(scater)
+library(scuttle)
 set.seed(1)
 counts <- counts(mockSCE())