+++ NEWS.md
@@ -1,5 +1,12 @@
 # DEVELOPMENT VERSION
 
+## Version 1.23.2 (2023-04-05)
+
+* Replace set.seed() with withr::with_seed()
+* Switch scater and scuttle dependencies. scater is now suggested which saves
+  a few dependencies
+* Move ggplot2 to suggested for the same reason
+
 ## Version 1.23.1 (2023-01-30)
 
 * Fix bug in splatPopSimulate() where conditional group assignments were